Adding Keyfile check

+++ SyncTool/FileHandler.cs
@@ -48,8 +48,9 @@ namespace SyncTool
 
         public static void DeleteFile(PBO pbo)
         {
-
-            File.Delete(pbo.filePath + pbo.fileName);
+            //Delete only if the path has a keyfile and is not a keyfile
+            if (File.Exists(pbo.filePath + KeyFile) && !(pbo.filePath + KeyFile).Equals(pbo.filePath + pbo.fileName))
+                File.Delete(pbo.filePath + pbo.fileName);
 
         }
